Added files via upload
This is a PHP script I wrote with the help of PHPexcel. This reads an Excel spreadsheet and uploads values. I wrote this script to make the process of uploading results to graph a simple approach. Script reads through then uploads results to the database which can then be maniluptated to display data. Please note the date is set to a static value while I continue working on this script.