Added files via upload
A Customer survey reporting system I have been working in. This project makes use of PHPExcel and JPgraph. The aim of this project is to better understand data in form of a graph which makes it easier for Management & Staff to understand performance. This is a work in progress currently with full functionality of reporting being worked on.